# src/main/java/com/pragma/powerup/application/dto/request/AuthenticationRequestDto.java
package com.pragma.powerup.application.dto.request;

import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
public class AuthenticationRequestDto {

    private String email;
    private String password;
}